move scrollbar state into TableInteractionState
